requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.8.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
tiktoken==0.5.2
tenacity==8.2.3
